requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
filelock==3.15.1
Flask==3.0.3
fsspec==2024.6.0
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.1
pillow==10.3.0
requests==2.32.3
sympy==1.12.1
torch==2.3.1+cpu
torchvision==0.18.1+cpu
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
